Fixes add column button
